Adds parentheses balancer
I found that the bot was sending a lot of tweets with orphan parentheses. So
I built a little utility that ensures that if the Markov chain creates a string
containing parentheses, that you end up with a single matched set.

Note that this is not a real paren balancer - don't use it for anything where
you actually need to balance them for real.

This changeset also introduces a utility function file.